Optimisé

+++ TP2/Ex5.c
@@ -2,7 +2,7 @@
 
 int main() {
     long int s = 0;
-    long long int p = 1;
+    double p = 1;
 
     for (int i = 1; i<=100; i++) {
         printf("%d\n", p);
@@ -10,5 +10,5 @@ int main() {
         p*=i;
     }
     printf("La somme des 100 premiers nombres positifs est : %ld\n\n", s);
-    printf("Le produit des 100 premier nombres positifs est : %lld", p);
+    printf("Le produit des 100 premier nombres positifs est : %lf", p);
 }
